jcmt2caom2ingest: allow for SEEINGST to be an empty string.
This bug is listed in scicom ticket 481
(https://trac.jach.hawaii.edu/trac/scicom/ticket/481).
This fix will cause no value to be recorded in CAOM if the
SEEINGST header has an empty string.

This fix only affects this specific header: its possible that other
fits header values are sometimes left as an empty string when the
software expects something it can format as a number. In that case
this fix should probably be included for all numeric header values
that are optional.