For remove_dict, check also old recipe instances
This is a temporary kludge for the problem that the "remove_dict"
considers planes to be from a different job -- and therefore
ineligible for removal -- if they have a different "run ID".  However
all the run IDs do change when the jobs are re-processed in the new
JSA processing system at JAC.  This prevents wrongly named old planes
from being removed.

With this commit, a file of the mappings between the two sets of job
numbers is read and used to check both IDs for each job.

A wider problem is that the JSA also has a number of products which are
duplicated between processing jobs.  This isn't really allowed (a CAOM-2
plane can only have one provenance and therefore only one "run ID") and
we should try to eliminate this, but the ingestion code might have to
try to cope with it in the meantime?