Extract get_project_pi_title function
This function, which gets the PI and title of a project, has been
extracted from jcmt2caom2.jcmt2caom2ingest and jcmt2caom2.raw where
the same kind of logic was duplicated.  This has been done so that
a necessary change -- checking that the OMP information hasn't
been obfuscated -- can be made in only one place.