Revert "Clear artifacts from raw planes being re-ingested"
This reverts commit af5e82b8598f920d37cd97d1f720cc7fa6267a4c.

This clean-up code is not currently necessary because jcmt2caom2.raw
always deletes all existing raw planes from the observation before
re-ingesting them.  Removing it because it causes artifacts in the
other hybrid subsystem to be removed.  Ideally jcmt2caom2.raw would
update rather than deleting and re-creating raw planes, in which
something like this would be required, but it would need to be updated
to be aware of hybrid mode.