Use guessed OBSID to optimize OSBnnn TAP query
The TAP query which looks up the OBS1 ... OBSCNT headers (which are
OBSIDSS values) in CAOM-2 used a LIKE expression to match a pattern
extracted from them against the OBSID values used as observation
identifiers.  This turns out to be too slow so this commit replaces
it with an equality expression based on the guessed OBSID value.