Set WCS information for catalog files
This uses Healpy to generate the corners of the tile, using the
TILENUM header to obtain the nested numbering scheme pixel number.
For the spectral WCS it uses the routine previously extracted from
the raw data ingestion code.

A new data structure has been added to store the WCS information
to avoid having to try to incorporate it into the existing
structures.