Remove "obsid_solitary" test for OBSnnn headers
Now that these are queried with an exact "guessed" obsid rather than a
LIKE expression, we no longer need to check that only one obsid is
returned.  Also remove the "earliest_obs" which was only used for an
obsolete logging feature and looks like it wouldn't work in the OBSnnn
case anyway.