jcmt2caom2ingest: simplify remove_dict handling
Now perform a TAP query which simply selects the planes which match
the current (or previous system) run ID.  Then take advantage of the
new CAOM-2 repository wrapper "allow_remove" option so that removing
all the planes from an observation automatically removes the whole
observation.  This prevents the ingestion code from having to try
to make that decision based on TAP information (which may be out of
date due to the delay present in "harvesting" at CADC).