Set metrics.sourceNumberDensity for catalog planes
This sets the "sourceNumberDensity" to 0 when a "tile-moc" is encountered,
or f[1].header['NAXIS2'] for "extent-cat" and "peak-cat".  Then
"build_metadict" retains the non-zero value.  The aim is to set the
"sourceNumberDensity" to zero if there are no sources detected (only a
"tile-moc" file is present) or the number of sources per tile otherwise.