Directly return "plane_dict" from "build_dict"
This adds a class "OrderedStrDict" which enforces the constraint that
values be strings, as was previously done by the "add_to_plane_dict"
method.  This class is like the typed classes of the "caom2.util.caom2_util"
module (but unfortunately none of those are suitable for use here).
We can then directly add values to the dictionary rather than needing to
call "add_to_plane_dict".

The "build_dict" method now constructs an instance of "OrderedStrDict"
and returns it rather than communicating through the member attribute
"self.plane_dict".

This commit retains the (slightly confusing) "plane_dict" nomenclature
to make the changes more straightforward.