Directly return "fitsuri_dict" from "build_dict"
The "build_dict" method now returns a "FileInfo" namedtuple instance
which contains the plane and "fitsuri" dictionaries.  The "fitsuri"
dictionary has been split into plain and custom parts.

This makes use of a new class "OrderedDefaultDict", which combines
the behavior of the "defaultdict" and "OrderedDict" classes, to
contain the "fitsuri" data.  This is set up so that, for each
"fitsuri" referenced, an "OrderedStrDict" is automatically created.