Simplify determination of part product type
Now store the part product types for each standard product as a
dictionary rather than a string which we will later have to parse.
If the PRODTYPE FITS header is present, the product type for each
extension is read directly from it (but we only look at the primary
HDU and first extension) and other parts are assumed to be of type
"auxiliary".  We then use caom2.caom2_enums.ProductType to validate
the product types we are setting rather than having to list the
valid types in a regular expression.  The "ngs_prepare_files.py"
example script's usage of the PRODTYPE header has been updated.