Set sourceNumberDensity per square degree
CADC have confirmed that the metrics.sourceNumberDensity field should
be in units of sources per square degree.  This commit therefore divides
the number of sources by the area of coverage in the JSA tile.  It does
this by getting the area from the "tile-moc" file.  The area and number
of sources are written into the new "custom" plane information.  This
allows the source number density to be calculated when preparing the
fits2caom2 override file.  Note that when ingestion the "peak" plane,
we have to get the tile area from the "extent" plane, since that is
where the "tile-moc" is.