Set rsp/rimg product type to auxiliary
Product type(s) "preview" (and "thumbnail") are apparently now used
to identify the images to be show in the Advanced Search interface,
replacing automatic determination of the image names from the
observation and plane identifiers.  This commit therefore sets
the product type for "rsp" and "rimg" files to something other
than "preview" -- namely "auxiliary" as no better alternative
exists.