Add previews as CAOM-2 artifacts
Now look for PNG as well as FITS files.  PNG files are identified
as beloning to an observation and plane by their jsa:obsid/jsa:asn_id
and jsa:productID EXIF keywords.  The 256 pixel file is used as the
CAOM-2 "thumbnail" and the 1024 pixel file as the "preview".  If
previews are being added to a "reduced" plane and the observation
has an equivalent "raw" plane, then the previews are used for it
too.