ingest: update SEEINGST special handling
Previously it would fail if the SEEEINGST was an all white space
string, and I think if it was an empty string. This should handle both
cases.  I also updated the NB comment so it was correct (an empty
string would fail the > 0.0 test with a can't compare float to string
issue).